Refactor test helper functions
Moved helper functions from main test file into their own test file to
improve legibility.

Resolves ##174542544.
